# specs/001-architect-client-portal/contracts/api.yaml
openapi: "3.1.0"
info:
  title: ArchiDrive API
  version: "1.0.0"
  description: Architect-Client Design Portal — REST + WebSocket API
servers:
  - url: http://localhost:8000
    description: Development
security:
  - bearerAuth: []
paths:
...
  /api/projects:
    get:
      tags: [Projects]
      summary: List current user's projects
      parameters:
        - in: query
          name: archived
          schema: { type: boolean }
          description: Filter by archive status
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectResponse"
    post:
      tags: [Projects]
      summary: Create a new project (architect only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 255 }
                description: { type: string }
                firm_id: { type: integer, nullable: true }
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
  /api/projects/{project_id}:
    get:
      tags: [Projects]
      summary: Get project details
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetailResponse"
        "404": { description: Project not found or access denied }
    patch:
      tags: [Projects]
      summary: Update project (architect only)
      responses:
        "200": { description: Updated }
    delete:
      tags: [Projects]
      summary: Delete/archive project (architect only)
      parameters:
        - in: query
          name: archive_only
          schema: { type: boolean, default: true }
      responses:
        "204": { description: Deleted/archived }
  /api/projects/{project_id}/updates:
    get:
      tags: [Projects]
      summary: Poll for project updates (WebSocket fallback)
      parameters:
        - in: query
          name: since
          schema: { type: string, format: date-time }
          description: ISO 8601 timestamp
      responses:
        "200":
          description: Update check
          content:
            application/json:
              schema:
                type: object
                properties:
                  has_updates: { type: boolean }
                  timestamp: { type: string }

  # ── Invitations ───────────────────────────────────────────
  /api/projects/{project_id}/invite:
    post:
      tags: [Invitations]
      summary: Invite a client to a project (architect only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "201":
          description: Invitation sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationResponse"
...
  /api/projects/{project_id}/milestones:
    get:
      tags: [Milestones]
      summary: List project milestones (ordered by position)
      responses:
        "200":
          description: Milestone list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MilestoneResponse"
    post:
      tags: [Milestones]
      summary: Create a milestone (architect only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 255 }
                description: { type: string }
                position: { type: integer }
      responses:
        "201":
          description: Milestone created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MilestoneResponse"
  /api/milestones/{milestone_id}:
    patch:
      tags: [Milestones]
      summary: Update milestone (architect only)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                position: { type: integer }
                is_completed: { type: boolean }
      responses:
        "200":
          description: Updated milestone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MilestoneResponse"
    delete:
      tags: [Milestones]
      summary: Delete milestone (architect only)
      responses:
        "204": { description: Deleted }

  # ── Comments ──────────────────────────────────────────────
...
  /ws/projects/{project_id}:
    get:
      tags: [WebSocket]
      summary: Real-time project updates
      description: |
        WebSocket endpoint for real-time project updates.
        Connect with: `ws://host/ws/projects/{project_id}?token=<jwt>`
        
        Server → Client events:
        - `{"type": "file_uploaded", "file_id": "uuid", "filename": "..."}`
        - `{"type": "file_deleted", "file_id": "uuid"}`
        - `{"type": "file_updated", "file_id": "uuid", "filename": "..."}`
        - `{"type": "milestone_updated", "milestone_id": 1, "is_completed": true}`
        - `{"type": "comment_added", "file_id": "uuid", "comment_id": 1}`
        - `{"type": "ping"}`
        
        Client → Server events:
        - `{"type": "ping"}` → server responds with `{"type": "pong"}`
      parameters:
        - in: path
          name: project_id
          required: true
          schema: { type: integer }
        - in: query
          name: token
          required: true
          schema: { type: string }
          description: JWT access token
      responses:
        "101": { description: WebSocket upgrade successful }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserResponse:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        name: { type: string }
        role: { type: string, enum: [architect, client] }
        firm_id: { type: integer, nullable: true }
        is_firm_admin: { type: boolean }
        is_verified: { type: boolean }
        created_at: { type: string, format: date-time }
...
    ProjectResponse:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string, nullable: true }
        owner_id: { type: integer }
        firm_id: { type: integer, nullable: true }
        is_archived: { type: boolean }
        file_count: { type: integer }
        milestone_count: { type: integer }
        completed_milestone_count: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ProjectDetailResponse:
      allOf:
        - $ref: "#/components/schemas/ProjectResponse"
        - type: object
          properties:
            milestones:
              type: array
              items:
                $ref: "#/components/schemas/MilestoneResponse"
            files:
              type: array
              items:
                $ref: "#/components/schemas/DesignFileResponse"
    InvitationResponse:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        token: { type: string }
        expires_at: { type: string, format: date-time }
        is_used: { type: boolean }
        created_at: { type: string, format: date-time }
    DesignFileResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        project_id: { type: integer }
        milestone_id: { type: integer, nullable: true }
        filename: { type: string }
        file_type: { type: string }
        content_type: { type: string }
        file_size: { type: integer }
        thumbnail_status:
          type: string
          enum: [pending, processing, complete, failed, unsupported]
        preview_status: { type: string, nullable: true }
        is_deleted: { type: boolean }
        version_number: { type: integer }
        comment_count: { type: integer }
        uploaded_by:
          $ref: "#/components/schemas/UserResponse"
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    MilestoneResponse:
      type: object
      properties:
        id: { type: integer }
        project_id: { type: integer }
        name: { type: string }
        description: { type: string, nullable: true }
        position: { type: integer }
        is_completed: { type: boolean }
        completed_at: { type: string, format: date-time, nullable: true }
        file_count: { type: integer }
        created_at: { type: string, format: date-time }
...